+++ src/gamestate/TileMap.h
@@ -78,7 +78,7 @@ private:
     void drawTexturedBackground(TileMapLayer& layer, const double& x, const double& y, std::shared_ptr<GameView>& view);
     void setTileDestructibleEventFlag(std::shared_ptr<LayerTile>& tile, const uint& x, const uint& y,
         const TileDestructType& type, const quint16& extraByte);
-    bool advanceDestructibleTileAnimation(std::shared_ptr<LayerTile>& tile, const double& x, const double& y, const QString& soundName);
+    bool advanceDestructibleTileAnimation(std::shared_ptr<LayerTile>& tile, const int& x, const int& y, const QString& soundName);
     std::shared_ptr<LayerTile> cloneDefaultLayerTile(int x, int y);
     std::unique_ptr<Tileset> levelTileset;
     QVector<TileMapLayer> levelLayout;
